ml-service/requirements.txt
flask
flask-cors
scikit-learn
scikit-surprise
numpy<2
pandas
joblib
pymongo
python-dotenv
